Solved an almost infinite loop bug in query optimization.

+++ core/search.h
@@ -41,7 +41,7 @@ namespace search
 			void sort(vector<condition**> list, condition_sort_logic logic, watable& wat)
 			{
 				bool test = false;
-				for(int i = 0; i < list.size() - 1; i++)
+				for(int i = 0; i < (((int)list.size()) - 1); i++)
 				{
 					for(int j = i + 1; j < list.size(); j++)
 					{
